compile ze_common with float.

+++ include/ze/common/types.h
@@ -16,7 +16,7 @@ using uint8_t   = std::uint8_t;
 using uint16_t  = std::uint16_t;
 using uint32_t  = std::uint32_t;
 using uint64_t  = std::uint64_t;
-using FloatType = double;
+using FloatType = float;
 
 //------------------------------------------------------------------------------
 // Typedefs of commonly used Eigen matrices and vectors.
